Move to utils and add make_cog option
